Minor fixup: rename oauth_flow_manager module
Should always have been oauth2_flow_manager, for consistency with the
rest of the OAuth2 modules.
